requirements.txt
Flask==1.0.2
PyYaml==3.12
SQLAlchemy==1.2.8
Flask-SQLAlchemy==2.3.2
mysqlclient==1.3.12